Adding the first version of the game
Desktop version of a game to learn playing piano. This is a simple html file with an interactive piano. The game consists on pressing the key that appears in the top of the screen.
Next steps: correct the name of the black keys, reproduce sound, recognize the played key, connect a real keyboard and recognize the pressed key (midi experimental api)